streamlit_app/requirements.txt
streamlit>=1.30.0
psycopg2-binary>=2.9.9
pandas>=2.0.0
geopandas>=0.14.0
pydeck>=0.8.0
folium>=0.15.0
streamlit-folium>=0.15.0
plotly>=5.18.0
shapely>=2.0.0
